+++ client/Packages/com.beamable/Runtime/Modules/Console/ConsoleFlow.cs
@@ -118,7 +118,6 @@ namespace Beamable.Console
 
 		private void Update()
 		{
-			var _ = BeamContext.ForPlayer(_playerCode).OnReady;
 			if (!_isInitialized) return;
 
 			if (_showNextTick)